split the worker, reloader proxy and process objects apart
It was convenient to share the same object but splitting it apart makes
it much more clear what methods may be called from the parent (new
Worker class) and which methods are part of the reloader proxy.

The Worker class now acts as a helper wrapping the worker process and
starting/stopping it without any leaks.