copy sys.path to the worker process and ensure hupper is on the PYTHONPATH
zc.buildout directly enhances sys.path of a process and thus the
path is not inherited by any children. This fix puts hupper on the
PYTHONPATH such that it is importable in the worker process. Finally it
copies the entire sys.path to the worker to restore any virtualenv-like
path settings that existed in the parent.

fixes #25